+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [HexaPDF::Layout::Style#box_options] for specifying box initialization options
 * [HexaPDF::Layout::Style#font_bold] and [HexaPDF::Layout::Style#font_italic]
   for setting bold and/or italic variants independently of the font name
+* [HexaPDF::PDFArray#map!] for mapping elements in-place
 
 ### Changed